Comparator hotfix - 32-bit values now unsigned

+++ Components/Cmp/CmpBase.cs
@@ -11,8 +11,8 @@ namespace IntegratedCircuits.Components.Cmp
         }
         protected override void CircuitLogicUpdate()
         {
-            int a = Util.ReadIntFromInputs(Inputs, 0, bits - 1);
-            int b = Util.ReadIntFromInputs(Inputs, bits, 2 * bits - 1);
+            uint a = (uint)Util.ReadIntFromInputs(Inputs, 0, bits - 1);
+            uint b = (uint)Util.ReadIntFromInputs(Inputs, bits, 2 * bits - 1);
             Outputs[0].On = a < b;
             Outputs[1].On = a == b;
             Outputs[2].On = a > b;
